Time: 1 ms (100%), Space: 48.9 MB (33.06%) - LeetHub

+++ 0004-median-of-two-sorted-arrays/0004-median-of-two-sorted-arrays.java
@@ -1,30 +1,29 @@
 class Solution {
     public double findMedianSortedArrays(int[] nums1, int[] nums2) {
-        int n1 = nums1.length, n2 = nums2.length;
-        if (n1 > n2)
-            return findMedianSortedArrays(nums2, nums1);
-        int low = 0, high = n1;
-        while (low <= high) {
-            int cut1 = low + high >> 1;
-            int cut2 = (n1 + n2 + 1) / 2 - cut1;
+        int n1=nums1.length,n2=nums2.length;
+        if(n1>n2) return findMedianSortedArrays(nums2,nums1);
+        int low=0,high=n1;
+        while(low<=high){
+            int cut1=low+high>>1;
+            int cut2=(n1+n2+1)/2-cut1;
 
-            int l1 = cut1 == 0 ? Integer.MIN_VALUE : nums1[cut1 - 1];
-            int l2 = cut2 == 0 ? Integer.MIN_VALUE : nums2[cut2 - 1];
+            int l1=cut1==0?Integer.MIN_VALUE:nums1[cut1-1];
+            int l2=cut2==0?Integer.MIN_VALUE:nums2[cut2-1];
 
-            int r1 = cut1 == n1 ? Integer.MAX_VALUE : nums1[cut1];
-            int r2 = cut2 == n2? Integer.MAX_VALUE : nums2[cut2];
+            int r1=cut1==n1?Integer.MAX_VALUE:nums1[cut1];
+            int r2=cut2==n2?Integer.MAX_VALUE:nums2[cut2];
 
-            if (l1 <= r2 && l2<=r1) {
-                if ((n1 + n2) % 2 == 0) {
-                    return (double) (Math.max(l1, l2) + Math.min(r1, r2))/2.0;
-                } else {
-                    return (double) Math.max(l1, l2);
+            if(l1<=r2&&l2<=r1){
+                if((n1+n2)%2==0){
+                    return (Math.max(l1,l2)+Math.min(r1,r2))/2.0;
+                }else{
+                    return Math.max(l1,l2);
                 }
             }
-            if (l1 > r2) {
-                high = cut1 - 1;
-            } else {
-                low = cut1 + 1;
+            if(l1>r2){
+                high=cut1-1;
+            }else{
+                low=cut1+1;
             }
         }
         return 0.0;
